change float to double

+++ DosingApp/DosingApp/Models/ApplicatorTank.cs
@@ -9,7 +9,7 @@ namespace DosingApp.Models
         public int ApplicatorTankId { get; set; }
 
         public string Name { get; set; }
-        public float? Volume { get; set; }
+        public double? Volume { get; set; }
 
         public int? ApplicatorId { get; set; }
         public virtual Applicator Applicator { get; set; }
