+++ DosingApp/DosingApp/Models/Recipe.cs
@@ -19,7 +19,7 @@ namespace DosingApp.Models
         public virtual Component Carrier { get; set; }
 
         public string Note { get; set; }
-        public float? CarrierReserve { get; set; }
+        public double? CarrierReserve { get; set; }
 
         public ICollection<RecipeComponent> RecipeComponents { get; set; }
     }
