+++ DosingApp/DosingApp/Models/TransportTank.cs
@@ -9,7 +9,7 @@ namespace DosingApp.Models
         public int TransportTankId { get; set; }
 
         public string Name { get; set; }
-        public float? Volume { get; set; }
+        public double? Volume { get; set; }
 
         public int? TransportId { get; set; }
         public virtual Transport Transport { get; set; }
